refactor: move files related to Simple Table
