+++ giraffe/src/components/SimpleTable/index.ts
@@ -1,0 +1,1 @@
+export {SimpleTableLayer} from './SimpleTableLayer'
