+++ giraffe/src/components/SizedTable.tsx
@@ -14,7 +14,7 @@ import {newTableFromConfig} from '../utils/newTable'
 import {RawFluxDataTable} from './RawFluxDataTable'
 import {FluxTablesTransform} from './FluxTablesTransform'
 import {TableGraphLayer} from './TableGraphLayer'
-import {SimpleTableLayer} from './SimpleTableLayer'
+import {SimpleTableLayer} from './SimpleTable'
 
 import {usePlotEnv} from '../utils/usePlotEnv'
